backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.11
python-dotenv==1.1.0
httpx==0.28.1
python-multipart==0.0.20
anthropic==0.52.0
openpyxl==3.1.5
xlrd==2.0.1
pytest==8.3.4
